# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Analizcim API",
    "description": "Excel analiz uygulaması için REST API. Bu API, satış ve alış dosyalarını analiz etmek, geçmişi yönetmek, tahminler oluşturmak ve kullanıcı yönetimi için kullanılır.",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "support@analizcim.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Yerel geliştirme sunucusu"
    }
  ],
...
  "paths": {
...
    "/user/preferences": {
      "get": {
        "tags": ["Kullanıcı Tercihleri"],
        "summary": "Kullanıcı tercihlerini getir",
        "operationId": "getUserPreferences",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "parameters": [
          {
            "name": "keys",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "İstenen tercik anahtarları (virgülle ayrılmış)"
          }
        ],
        "responses": {
          "200": {
            "description": "Kullanıcı tercihleri"
          }
        }
      },
      "put": {
        "tags": ["Kullanıcı Tercihleri"],
        "summary": "Kullanıcı tercihlerini güncelle",
        "operationId": "updateUserPreferences",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": {
                    "type": "string",
                    "enum": ["light", "dark", "system"]
                  },
                  "predictions_layout_id": {
                    "type": "string"
                  },
                  "predictions_card_order": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tercihler güncellendi"
          }
        }
      }
    },
    "/user/preferences/migrate": {
      "post": {
        "tags": ["Kullanıcı Tercihleri"],
        "summary": "Kullanıcı tercihlerini taşı",
        "operationId": "migrateUserPreferences",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": {
                    "type": "string"
                  },
                  "predictions_layout_id": {
                    "type": "string"
                  },
                  "predictions_card_order": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tercihler taşındı"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Oturum kimlik doğrulaması"
      },
      "adminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid",
        "description": "Admin yetkisi gerektiren işlemler için"
      }
    }
  }
}